work on common routes
