+++ packages/server/lib/project-base.ts
@@ -30,7 +30,7 @@ import Watchers from './watchers'
 import devServer from './plugins/dev-server'
 import preprocessor from './plugins/preprocessor'
 import { SpecsStore } from './specs-store'
-import { createRoutes as createE2ERoutes } from './routes'
+import { createRoutes as createE2ERoutes } from './routes-e2e'
 import { checkSupportFile } from './project_utils'
 import type { LaunchArgs } from './open_project'
 import { createRoutes as createRoutesCT } from './routes-ct'
